Non-working base build of a Dockerized Quirkafleeg
([If you have to ask...](https://twitter.com/gilesdring/status/591196432390029312))

This is **HORRIBLE** and breaks all good practice for Docker use, as far as I can tell.
I'm mainly doing this to learn Docker.

Currently creates a base build, but can't keep `signon`, `www`, `asset-manager`, `search`
up. Seems to be a `mysql` config issue. Probably password, or user, or something.

It's also a mahoosive image, with `mysql` and `mongo` bundled in. These need to be moved
out to separate containers. Let me get this working first!